+++ src/haversine.rs
@@ -1,7 +1,7 @@
 use crate::Point;
 
-pub fn haversine_distance(fix1: &dyn Point, fix2: &dyn Point) -> f64 {
-    const R: f64 = 6371.; // kilometres
+pub fn haversine_distance(fix1: &dyn Point, fix2: &dyn Point) -> f32 {
+    const R: f32 = 6371.; // kilometres
 
     let phi1 = fix1.latitude().to_radians();
     let phi2 = fix2.latitude().to_radians();
